+++ plugins/idea/src/main/kotlin/org/pastalab/fray/idea/ui/Colors.kt
@@ -7,7 +7,7 @@ object Colors {
   val THREAD_ENABLED_COLOR = JBColor(Color(207, 243, 212), Color(43, 136, 55))
   val THREAD_DISABLED_COLOR = JBColor(Color(238, 238, 240), Color(196, 198, 198))
 
-  fun getThreadColor(index: Long): JBColor {
+  fun getThreadColor(index: Int): JBColor {
     // Generate a consistent color based on thread index
     // Using HSB color model to create visually distinct colors
     val hue = ((index * 0.618033988749895) % 1).toFloat()
